added Oauth login functionto utils file for better readability
