+++ client/src/components/OAuth.jsx
@@ -1,32 +1,11 @@
 import React from 'react';
+import { handleSignIn } from '../utils/utils.js';
 
 function OAuthSignIn() {
-    const handleSignIn = () => {
-        const oauth2Endpoint = 'https://accounts.google.com/o/oauth2/v2/auth';
-
-        const params = {
-            'client_id': '210179651380-g1s8fdb4enh5rifte5qnbra1dbuef6ht.apps.googleusercontent.com',
-            'redirect_uri': 'http://localhost:5173/callback',
-            'response_type': 'token',
-            'scope': 'https://www.googleapis.com/auth/drive.metadata.readonly https://www.googleapis.com/auth/calendar.readonly',
-            'include_granted_scopes': 'true',
-            'state': 'pass-through value'
-        };
-
-        // Build the OAuth URL with query parameters
-        const url = new URL(oauth2Endpoint);
-        Object.keys(params).forEach(key => {
-            url.searchParams.append(key, params[key]);
-        });
-
-        // Redirect to the OAuth endpoint
-        window.location.href = url.toString();
-    };
-
     return (
         <button onClick={handleSignIn}>
             <svg className="google-icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48" width="18px" height="18px">
-            
+
                 <path fill="#FFC107" d="M43.611,20.083H42V20H24v8h11.303c-1.649,4.657-6.08,8-11.303,8c-6.627,0-12-5.373-12-12c0-6.627,5.373-12,12-12c3.059,0,5.842,1.154,7.961,3.039l5.657-5.657C34.046,6.053,29.268,4,24,4C12.955,4,4,12.955,4,24c0,11.045,8.955,20,20,20c11.045,0,20-8.955,20-20C44,22.659,43.862,21.35,43.611,20.083z" />
                 <path fill="#FF3D00" d="M6.306,14.691l6.571,4.819C14.655,15.108,18.961,12,24,12c3.059,0,5.842,1.154,7.961,3.039l5.657-5.657C34.046,6.053,29.268,4,24,4C16.318,4,9.656,8.337,6.306,14.691z" />
                 <path fill="#4CAF50" d="M24,44c5.166,0,9.86-1.977,13.409-5.192l-6.19-5.238C29.211,35.091,26.715,36,24,36c-5.202,0-9.619-3.317-11.283-7.946l-6.522,5.025C9.505,39.556,16.227,44,24,44z" />
